+++ kubernetes-tests/src/test/java/io/fabric8/kubernetes/client/mock/SecretCrudTest.java
@@ -30,7 +30,6 @@ import java.util.Collections;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
@@ -82,7 +81,7 @@ public class SecretCrudTest {
     secret2 = client.secrets().inNamespace("ns2").withName("secret2")
         .edit(s -> new SecretBuilder(s).removeFromData("one").build());
     assertNotNull(secret2);
-    assertNull(secret2.getData());
+    assertTrue(secret2.getData().isEmpty());
   }
 
   @Test